Change MC years weight to float instead of int (#105)
* Change MC years weight to float instead of int

* use float value instead of double

* display only 3 digits for mc years weight

+++ src/solver/variable/storage/averagedata.h
@@ -82,8 +82,8 @@ namespace AllYears
 		mutable double allYears; // FIX MEEE - Remove the mutable as soon as possible
 		unsigned int nbYearsCapacity;
 
-		std::vector<int>    yearsWeight;
-		int                 yearsWeightSum;
+		std::vector<float>    yearsWeight;
+		float                 yearsWeightSum;
 
 	}; // class AverageData
